backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
pymongo
python-multipart
python-dotenv
openai
httpx
google-generativeai
